video 4 to 13
Summary :  Video 4 (Layout design)
Description : in this lecture , we make the layout for our input page by making containers and design them. (2:29 am)

Summary : Video 5 (Extract widget)
Description : in this lecture we extract the widget and name it RepeatContainerCode and we apply colors on our cards or containers that we make earlier. (2:51 am)

Summary : Video 6 (icon and text)
Description : we make the icon and text of male and female box (3:28 am)

Summary : video 7 (smart work)
Description : we make the 2 more files : one for card reusable code and other one for icon & text reusable code. (3:34 am)

Summary : video 8 (gesture detector)
Description : in this lecture we focus on the gesture detecture and also do a thing widget wrap (3:55 am)

Summary : video 9 (enum)
Description : in this lecture we make the enum for genders and access it in the same way (4:00 am)

Summary : video 10 (ternary operator)
Description :  we just simple the code of active and deactive and card color by just using ternary opertors (4:10 am)

Summary : video 11 (Function object)
Description :  we use the gesture detector with function use as an object that make the code more simple. (4:22 am)

Summary : video 12 (const file)
Description : we make a file where we only put the variables that we use anywhere in the project code files

Summary : video 13 (Height slider)
Description : we make a slider for height as we move the slider the height also change